backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "api"
version = "1.0.0"
description = "API para dialogar con el backend"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "Brian Malavenda", email = "brian.malavenda@gmail.com"}
]

dependencies = [
    "flask>=2.3.0",
    "flask-cors>=4.0.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "cryptography>=3.4.0",
    "python-docx>=1.1.0",
    "pydub>=0.25.0",
    "pika>=1.3.2",
    "gunicorn>=20.0.0",
    "fastapi==0.115.0",
    "uvicorn==0.34.0",
    "python-multipart==0.0.12",
    "python-jose==3.3.0",
    "passlib==1.7.4",
    "soundfile==0.12.1",
    "python-telegram-bot==20.3"
]

[project.optional-dependencies]
test = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
]
dev = [
    "pytest>=7.0.0",
    "black>=22.0.0",
]

[project.scripts]
api = "api.__main__:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
addopts = "-v --tb=short"

[tool.hatch.build.targets.wheel]
packages = ["api"]